+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## Updated
+- Use items endpoint for more item metadata fields (DR-3855)
+
 ## [1.1.0] 2025-08-28
 ## Updated
 - updated middleware file to redirect synthetic collections (DR-3750)